+++ vscode/sampleprj/packages/observables/README.md
@@ -1,3 +1,7 @@
 # observables for DeviceScript
 
 A tiny observable library for [DeviceScript](https://microsoft.github.io/devicescript/) package.
+
+## Acknoledgments
+
+Inspired from [rxjs](https://rxjs.dev/) and [micro-observables](https://github.com/mike-north/micro-observable).
